[collection-type] Move ExplicitRelationCollectionRector to CollectionType namespace
