+++ rules-tests/CodeQuality/Rector/Class_/ExplicitRelationCollectionRector/config/configured_rule.php
@@ -3,7 +3,7 @@
 declare(strict_types=1);
 
 use Rector\Config\RectorConfig;
-use Rector\Doctrine\CodeQuality\Rector\Class_\ExplicitRelationCollectionRector;
+use Rector\Doctrine\TypedCollections\Rector\Class_\ExplicitRelationCollectionRector;
 
 return static function (RectorConfig $rectorConfig): void {
     $rectorConfig->rule(ExplicitRelationCollectionRector::class);
