+++ rules-tests/TypedCollections/Rector/Class_/ExplicitRelationCollectionRector/Fixture/odm_document.php.inc
@@ -1,9 +1,6 @@
 <?php
 
-namespace Rector\Doctrine\Tests\CodeQuality\Rector\Class_\ExplicitRelationCollectionRector\Fixture;
-
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\ORM\Mapping\ManyToMany;
+namespace Rector\Doctrine\Tests\TypedCollections\Rector\Class_\ExplicitRelationCollectionRector\Fixture;
 
 /**
  * @\Doctrine\ODM\MongoDB\Mapping\Annotations\Document
@@ -20,10 +17,7 @@ class OdmDocument
 -----
 <?php
 
-namespace Rector\Doctrine\Tests\CodeQuality\Rector\Class_\ExplicitRelationCollectionRector\Fixture;
-
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\ORM\Mapping\ManyToMany;
+namespace Rector\Doctrine\Tests\TypedCollections\Rector\Class_\ExplicitRelationCollectionRector\Fixture;
 
 /**
  * @\Doctrine\ODM\MongoDB\Mapping\Annotations\Document
